+++ src/parsers/problem/DMOJProblemParser.ts
@@ -12,14 +12,14 @@ export class DMOJProblemParser extends Parser {
     const elem = htmlToElement(html);
     const task = new TaskBuilder('DMOJ').setUrl(url);
 
-    const titleParts = elem.querySelector('.problem-title h2').textContent.split(' - ');
+    const titleParts = elem.querySelector('.problem-title h2').textContent.split(/ - | — |(?<=Contest \d+ [A-Z#]\d+) /);
     if (titleParts.length === 1) {
       task.setName(titleParts[0]);
     } else {
       let contestName = titleParts[0];
       let taskName = titleParts.slice(1).join(' - ');
 
-      const problemMatches = /(Problem \d+|[A-Z]\d+)$/.exec(contestName);
+      const problemMatches = /(Problem \d+|[A-Z#]\d+)$/.exec(contestName);
       if (problemMatches !== null) {
         contestName = contestName.substr(0, problemMatches.index).trim();
         taskName = `${problemMatches[0]} - ${taskName}`;