+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Fixed the parsing of HackerEarth titles when parsing contest problems
+
 ## [2.31.0](https://github.com/jmerle/competitive-companion/releases/tag/2.31.0) (2021-11-27)
 - Added parsers for A2 Online Judge
 - Added a contest parser for Luogu