+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Fixed the parsing of DMOJ contest names
 - Fixed the parsing of HackerEarth titles when parsing contest problems
 
 ## [2.31.0](https://github.com/jmerle/competitive-companion/releases/tag/2.31.0) (2021-11-27)