+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.31.1](https://github.com/jmerle/competitive-companion/releases/tag/2.31.1) (2021-12-18)
 - Fixed the parsing of DMOJ contest names
 - Fixed the parsing of HackerEarth titles when parsing contest problems
 - Fixed the parsing of Library Checker problems